Fixed rendering of render target containers in gum tool by standardizing Alpha to IRenderableIpso.

+++ RenderingLibrary/Math/Geometry/LineRectangle.cs
@@ -200,7 +200,7 @@ namespace RenderingLibrary.Math.Geometry
             set;
         }
 
-        public float Alpha => Color.A;
+        public int Alpha => Color.A;
